requirements.txt
beautifulsoup4==4.11.1
blinker==1.4
bs4==0.0.1
cheroot==8.6.0
CherryPy==18.6.1
click==8.1.3
dnspython==2.2.1
eventlet==0.33.1
Flask==2.1.2
Flask-Mail==0.9.1
Flask-Script==2.0.6
Flask-WTF==1.0.1
greenlet==1.1.2
itsdangerous==2.1.2
jaraco.classes==3.2.1
jaraco.collections==3.5.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.8.0
Jinja2==3.1.2
MarkupSafe==2.1.1
more-itertools==8.13.0
portend==3.1.0
pyaml==21.10.1
pytz==2022.1
PyYAML==6.0
shelljob==0.6.3
six==1.16.0
soupsieve==2.3.2.post1
tempora==5.0.1
Werkzeug==2.1.2
WTForms==3.0.1
zc.lockfile==2.0
